+++ .eslintrc.js
@@ -38,6 +38,13 @@ module.exports = {
           'object',
           'type',
         ],
+        pathGroups: [
+          {
+            pattern: '~/**',
+            group: 'parent',
+            position: 'before',
+          },
+        ],
         alphabetize: {
           order: 'asc',
         },
