+++ src/components/features/Article/ArticleContent.tsx
@@ -4,14 +4,15 @@ import isEqual from 'react-fast-compare';
 import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 
+import { TextTagList } from '~/components/molecules';
+import { ProfileCard } from '~/components/organisms';
+
 import { CodeBlock } from './CodeBlock';
 import { ImageBlock } from './ImageBlock';
 import { LinkBlock } from './LinkBlock';
 import { TableBlock } from './TableBlock';
 
 import { css } from 'styled-system/css';
-import { TextTagList } from '~/components/molecules';
-import { ProfileCard } from '~/components/organisms';
 
 type Props = {
   title: string;
