+++ src/components/organisms/ArticleSummaryCard/index.tsx
@@ -2,8 +2,9 @@ import { format } from 'date-fns';
 import React, { useMemo } from 'react';
 import isEqual from 'react-fast-compare';
 
-import { css } from 'styled-system/css';
 import { TextTagList } from '~/components/molecules';
+
+import { css } from 'styled-system/css';
 
 type Props = {
   slug: string;
