+++ src/components/organisms/PageHeader/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import isEqual from 'react-fast-compare';
 
-import { PageHeaderView } from './PageHeaderView';
-
 import { hrefMap } from '~/consts';
+
+import { PageHeaderView } from './PageHeaderView';
 
 export const PageHeader = React.memo(
   () => <PageHeaderView hrefToHome={hrefMap.home} />,
