+++ src/lib/isColorMode.ts
@@ -1,6 +1,5 @@
-import type { ColorMode } from '~/types';
-
 import { colorModeMap } from '~/consts';
+import type { ColorMode } from '~/types';
 
 export const isColorMode = (value: unknown): value is ColorMode =>
   Object.keys(colorModeMap).some(
